Fixed comments: Renamed variables, moved initialParams to utils; Added error handling for fetchMovies
